+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @kehto/demo
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @kehto/runtime@0.8.0
+  - @kehto/services@0.8.0
+  - @kehto/shell@0.8.0
+
 ## 0.0.7
 
 ### Patch Changes